Split code into modules
